Refactor ChaCha20BaseIntrinsics
